# server/internal/handlers/auth.go
package handlers

import (
	"net/http"

	"github.com/Hana-ame/chat-app/server/internal/auth"
	"github.com/Hana-ame/chat-app/server/internal/logutil"
)

type registerReq struct {
	Email    string `json:"email"`
	Username string `json:"username"`
	Password string `json:"password"`
}

type loginReq struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

type sessionResp struct {
	User        any    `json:"user"`
	AccessToken string `json:"access_token"`
	ExpiresIn   int64  `json:"expires_in"`
}

// Register godoc
// @Summary      Register a new user
// @Description  Create account with email, username and password
// @Tags         auth
// @Param        body  body  registerReq  true  "Registration details"
// @Success      200  {object}  sessionResp
// @Failure      409  {object}  map[string]any
// @Router       /api/auth/register [post]
func (s *Server) Register(w http.ResponseWriter, r *http.Request) {
	var req registerReq
	if err := decodeJSON(r, &req); err != nil {
		writeError(w, http.StatusBadRequest, "bad_request", err.Error())
		return
	}
	email := auth.NormalizeEmail(req.Email)
	username, err := auth.ValidateUsername(req.Username)
	if err != nil {
		writeError(w, http.StatusBadRequest, "invalid_username", err.Error())
		return
	}
	hash, err := auth.HashPassword(req.Password)
	if err != nil {
		writeError(w, http.StatusBadRequest, "weak_password", err.Error())
		return
	}
	u, err := s.Services.User.Create(r.Context(), email, username, hash)
	if err != nil {
		status, code := mapServiceError(err)
		if status == http.StatusConflict {
			logutil.Warn("register conflict: email=%s username=%s", email, username)
			writeError(w, status, "already_taken", "email or username already taken")
			return
		}
		writeError(w, status, code, err.Error())
		return
	}
	logutil.Info("user registered: %s (username=%s email=%s)", logutil.SafeID(u.ID), username, email)
	s.issueSession(w, r, u.ID)
}

// Login godoc
// @Summary      Log in
// @Description  Authenticate with email and password
// @Tags         auth
// @Param        body  body  loginReq  true  "Login credentials"
// @Success      200  {object}  sessionResp
// @Failure      401  {object}  map[string]any
// @Router       /api/auth/login [post]
func (s *Server) Login(w http.ResponseWriter, r *http.Request) {
	var req loginReq
	if err := decodeJSON(r, &req); err != nil {
		writeError(w, http.StatusBadRequest, "bad_request", err.Error())
		return
	}
	email := auth.NormalizeEmail(req.Email)
	u, hash, err := s.Services.User.GetByEmail(r.Context(), email)
	if err != nil {
		status, code := mapServiceError(err)
		if status == http.StatusNotFound {
			logutil.Warn("login failed: email=%s (not found)", email)
			writeError(w, http.StatusUnauthorized, "invalid_credentials", "invalid email or password")
			return
		}
		writeError(w, status, code, err.Error())
		return
	}
	if err := auth.VerifyPassword(hash, req.Password); err != nil {
		logutil.Warn("login failed: email=%s (wrong password)", email)
		writeError(w, http.StatusUnauthorized, "invalid_credentials", "invalid email or password")
		return
	}
	logutil.Info("user logged in: %s (%s)", logutil.SafeID(u.ID), email)
	s.issueSession(w, r, u.ID)
}

func (s *Server) issueSession(w http.ResponseWriter, r *http.Request, userID string) {
	access, _, err := s.Auth.IssueAccessToken(userID)
	if err != nil {
		writeError(w, http.StatusInternalServerError, "internal", err.Error())
		return
	}
	raw, hash := auth.GenerateRefreshToken()
	if _, err := s.Services.CreateRefreshToken(r.Context(), userID, hash, s.Cfg.RefreshTokenTTL); err != nil {
		writeError(w, http.StatusInternalServerError, "internal", err.Error())
		return
	}
	u, err := s.Services.User.GetByID(r.Context(), userID)
	if err != nil {
		status, code := mapServiceError(err)
		writeError(w, status, code, err.Error())
		return
	}
	setAuthCookie(w, r, "access_token", access, "/", s.Cfg.AccessTokenTTL)
	setRefreshCookie(w, r, raw, s.Cfg.RefreshTokenTTL)
	writeJSON(w, http.StatusOK, sessionResp{
		User:        u,
		AccessToken: access,
		ExpiresIn:   int64(s.Cfg.AccessTokenTTL.Seconds()),
	})
}

// Refresh godoc
// @Summary      Refresh access token
// @Description  Exchange a refresh token for a new session (via httpOnly cookie)
// @Tags         auth
// @Success      200  {object}  sessionResp
// @Failure      401  {object}  map[string]any
// @Router       /api/auth/refresh [post]
func (s *Server) Refresh(w http.ResponseWriter, r *http.Request) {
	c, err := r.Cookie("refresh_token")
	if err != nil {
		writeError(w, http.StatusBadRequest, "bad_request", "refresh token missing")
		return
	}
	s.refreshMu.Lock()
	defer s.refreshMu.Unlock()
	hash := auth.HashRefreshToken(c.Value)
	rt, err := s.Services.FindAndDeleteRefreshToken(r.Context(), hash)
	if err != nil {
		writeError(w, http.StatusUnauthorized, "refresh_invalid", "invalid refresh token")
		return
	}
	if rt.ExpiresAt.Before(timeNow()) {
		writeError(w, http.StatusUnauthorized, "refresh_expired", "refresh token expired")
		return
	}
	s.issueSession(w, r, rt.UserID)
}

// Logout godoc
// @Summary      Log out
// @Description  Revoke all refresh tokens for the user and clear cookie
// @Tags         auth
// @Security     BearerAuth
// @Success      200  {object}  map[string]any
// @Router       /api/auth/logout [post]
func (s *Server) Logout(w http.ResponseWriter, r *http.Request) {
	u := userFrom(r.Context())
	if u == nil {
		writeError(w, http.StatusUnauthorized, "unauthorized", "missing user")
		return
	}
	s.refreshMu.Lock()
	err := s.Services.DeleteUserRefreshTokens(r.Context(), u.ID)
	s.refreshMu.Unlock()
	clearRefreshCookie(w, r)
	clearAccessTokenCookie(w, r)
	if err != nil {
		w.Header().Set("X-Error", err.Error())
	}
	logutil.Info("user logged out: %s", logutil.SafeID(u.ID))
	writeJSON(w, http.StatusOK, map[string]any{"ok": true})
}

// Me godoc
// @Summary      Get current user
// @Description  Return the authenticated user's profile
// @Tags         users
// @Security     BearerAuth
// @Success      200  {object}  models.User
// @Router       /api/users/me [get]
func (s *Server) Me(w http.ResponseWriter, r *http.Request) {
	u := userFrom(r.Context())
	if u == nil {
		writeError(w, http.StatusUnauthorized, "unauthorized", "")
		return
	}
	writeJSON(w, http.StatusOK, u)
}