# server/internal/handlers/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chat App API",
    "version": "0.9.3",
    "description": "实时聊天应用后端 API。\n\n认证方式:\n- `Authorization: Bearer <access_token>` 请求头（首选）\n- `access_token` Cookie（SSE 等 WebSocket 场景回退）\n- `refresh_token` Cookie（仅 `/api/auth/refresh`）\n\n所有错误响应统一格式:\n```json\n{ \"error\": \"error_code\", \"message\": \"human-readable message\" }\n```\n"
  },
  "servers": [
    {
      "url": "https://wsl-8080.moonchan.xyz",
      "description": "生产服务器"
    },
    {
      "url": "http://localhost:8080",
      "description": "本地开发"
    }
  ],
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "summary": "注册新用户",
        "operationId": "register",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "$ref": "#/components/headers/SetCookie"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "description": "用户已存在",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "example": {
                        "error": "already_taken",
                        "message": "email or username already taken"
                      }
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "登录",
        "operationId": "login",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "$ref": "#/components/headers/SetCookie"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "description": "凭据无效",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "example": {
                        "error": "invalid_credentials"
                      }
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "summary": "刷新 access token",
        "description": "从 Cookie 中读取 refresh_token，轮换机制（一次性）",
        "operationId": "refreshToken",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "刷新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "$ref": "#/components/headers/SetCookie"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "description": "refresh token 无效或过期",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "example": {
                        "error": "refresh_invalid"
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "登出",
        "description": "清除所有 refresh token",
        "operationId": "logout",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "登出成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/users/me": {
      "get": {
        "summary": "获取当前用户信息",
        "operationId": "getMe",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "patch": {
        "summary": "更新当前用户信息",
        "operationId": "updateMe",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "用户名已存在",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "example": {
                        "error": "username_taken"
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/upload": {
      "get": {
        "summary": "上传页面（HTML）",
        "operationId": "uploadPage",
        "responses": {
          "200": {
            "description": "上传 HTML 页面"
          }
        }
      },
      "put": {
        "summary": "裸 body 上传文件",
        "operationId": "uploadPut",
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "上传成功"
          },
          "413": {
            "description": "文件过大"
          }
        }
      },
      "post": {
        "summary": "multipart/form-data 上传文件",
        "operationId": "uploadPost",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "上传成功"
          },
          "413": {
            "description": "文件过大"
          }
        }
      }
    },
    "/api/upload/{path}": {
      "put": {
        "summary": "带文件名上传",
        "operationId": "uploadPutWithPath",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Content-Type",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "上传成功"
          }
        }
      }
    },
    "/api/local/{path}": {
      "get": {
        "summary": "下载已上传文件",
        "operationId": "downloadFile",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "delete",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "文件内容"
          },
          "404": {
            "description": "文件不存在"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "`Authorization: Bearer <access_token>`\naccess_token 有效期 15 分钟，通过 `/api/auth/refresh` 续期。\n"
      }
    },
    "headers": {
      "SetCookie": {
        "schema": {
          "type": "string"
        },
        "description": "设置 `access_token` 和 `refresh_token` Cookie。\nrefresh_token 为 httpOnly，用于 `/api/auth/refresh`。\n"
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                {
                  "example": {
                    "error": "bad_request",
                    "message": "invalid request body"
                  }
                }
              ]
            }
          }
        }
      },
      "Unauthorized": {
        "description": "未认证",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                {
                  "example": {
                    "error": "unauthorized",
                    "message": "missing or invalid access token"
                  }
                }
              ]
            }
          }
        }
      },
...
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                {
                  "example": {
                    "error": "not_found",
                    "message": "resource not found"
                  }
                }
              ]
            }
          }
        }
      },
      "RateLimited": {
        "description": "请求频率超限",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                {
                  "example": {
                    "error": "rate_limited",
                    "message": "too many requests"
                  }
                }
              ]
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "错误码（稳定，可用于客户端判断）",
            "example": "bad_request"
          },
          "message": {
            "type": "string",
            "description": "人类可读的错误描述",
            "example": "invalid request body"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": [
          "email",
          "username",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "username": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32,
            "example": "alice"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 128,
            "example": "securePassword123"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "securePassword123"
          }
        }
      },
      "SessionResponse": {
        "type": "object",
        "required": [
          "user",
          "access_token",
          "expires_in"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "access_token": {
            "type": "string",
            "description": "JWT access token"
          },
          "expires_in": {
            "type": "integer",
            "description": "过期时间（秒）",
            "example": 900
          }
        }
      },
      "User": {
        "type": "object",
        "required": [
          "id",
          "username",
          "avatar_color",
          "status",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "example": "abc123def456"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "username": {
            "type": "string",
            "example": "alice"
          },
          "avatar_color": {
            "type": "string",
            "description": "默认头像背景色（Hex）",
            "example": "#5865F2"
          },
          "avatar_url": {
            "type": "string",
            "format": "uri",
            "example": "https://upload.moonchan.xyz/api/xxx/avatar.png"
          },
          "status": {
            "type": "string",
            "enum": [
              "online",
              "offline",
              "away"
            ],
            "example": "online"
          },
          "role": {
            "type": "string",
            "example": "admin"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "avatar_color": {
            "type": "string",
            "example": "#5865F2"
          },
          "avatar_url": {
            "type": "string",
            "format": "uri"
          }
        }
      },
...
    }
  }
}